% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/authentication_functions.R
\name{authenticate}
\alias{authenticate}
\title{Authenticate with the UNGA-DM Database API}
\usage{
authenticate(username, password)
}
\arguments{
\item{username}{A string. Your UNGA-DM Database API username.}

\item{password}{A string. Your UNGA-DM Database API password.}
}
\value{
This function returns an object of class \code{ungadm_session}.
}
\description{
This function allows you to authenticate with the UNGA-DM
  Database API. It creates a session. You need to provide your username and
  password to authenticate. This function returns an object of class
  \code{ungadm_session} that contains the credentials necessary to request
  password-protected data from the UNGA-DM Database API. You need to pass
  this session object into any function that requires a \code{session}
  argument. If your session times out, you'll need to run this function again
  to create a new session.
}
\examples{
\dontrun{
session <- authenticate(
  username = "USERNAME",
  password = "PASSWORD"
)}

}
