% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/describe_functions.R
\name{describe_datasets}
\alias{describe_datasets}
\title{Describe the UNGA-DM datasets}
\usage{
describe_datasets(session)
}
\arguments{
\item{session}{An object of class \code{undata_app_session} created by
\code{authenticate()}.}
}
\value{
This function returns a tibble containing 5 variables. There is one
  observation per dataset in UNGA-DM Database.
}
\description{
This function provides descriptions of all of the datasets that
  are in the UNGA-DM Database.
}
\examples{
\dontrun{
session <- authenticate(
  username = "USERNAME",
  password = "PASSWORD"
)

out <- describe_datasets(
  session = session
)}

}
