% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/describe_functions.R
\name{describe_variables}
\alias{describe_variables}
\title{Describe variables in a UNGA-DM dataset}
\usage{
describe_variables(session, dataset)
}
\arguments{
\item{session}{An object of class \code{ungadm_session} created by
\code{authenticate()}.}

\item{dataset}{A string. The name of a dataset in the UNGA-DM Database. Run
\code{list_datasets()} to get a list of valid values.}
}
\value{
This function returns a tibble containing 7 variables. There is one
  observation per variable in the specified dataset.
}
\description{
This function provides descriptions of all of the variables in a
  dataset in the UNGA-DM Database. You have to specify a dataset.
}
\examples{
\dontrun{
session <- authenticate(
  username = "USERNAME",
  password = "PASSWORD"
)

out <- describe_variables(
  session = session,
  dataset = "decisions"
)}

}
