% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_data.R
\name{download_data}
\alias{download_data}
\title{Download data from the UNGA-DM Database}
\usage{
download_data(session, dataset, filters = NULL, variables = NULL)
}
\arguments{
\item{session}{An object of class \code{ungadm_session} created by
\code{authenticate()}.}

\item{dataset}{A string. The name of a dataset in UNGA-DM Database. Use
\code{list_datasets()} to get a list of valid values.}

\item{filters}{A named list. The default is \code{NULL}. Each element in the
list specifies a filter to apply to the data. The name of each element
should be the name of a variable in the specified dataset and the
corresponding value should be a value or vector of values that the variable
can take. The results will only include observations where the variable
equals one of the provided values. If you specify multiple filters, the
results will only include observations that match all of the filters. For
numeric variables, you can add \code{_min} or \code{_max} to the end of the
variable name to specify a minimum or maximum value.}

\item{variables}{A string vector. The default is \code{NULL}. The results
will only include the variables in the vector. Use \code{list_variables()}
to get a list of valid values. The function will throw an error if you
provide an invalid variable name.}
}
\value{
This function returns a tibble that contains the requested data.
}
\description{
This function allows you to download data from the UNGA-DM
  Database via the UNGA-DM Database API. You need to specify a dataset in the
  UNGA-DM Database. You can also specify filters and select specific
  variables. You can use \code{describe_variables()} to learn more about the
  variables available in each dataset.
}
\details{
The UNGA-DM Database API has a rate limit, so this function
  downloads data from the API in batches. It downloads 10,000 observations
  every 5 seconds. It prints a message to the console that indicates how many
  observations you've requested and approximately how long it will take to
  download the data. It also prints a message after every batch that
  indicates the current progress. After your download is complete, it will
  print the suggested citations for the database, the working paper that
  introduces the database, and the R package. Please use these citations if
  you use the data in a paper.
}
\examples{
\dontrun{
session <- authenticate(
  username = "USERNAME",
  password = "PASSWORD"
)

out <- download_data(
  session = session,
  dataset = "decisions",
  filters = list(
    decision_mode = c("recorded vote", "non-recorded vote")
  )
)

out <- download_data(
  session = session,
  dataset = "decisions",
  filters = list(
    decision_mode = "recorded_vote",
    meeting_date_min = "2018-08-01"
  ),
  variables = c("decision_id", "meeting_record_id", "topic")
)}

}
