% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/list_functions.R
\name{list_datasets}
\alias{list_datasets}
\title{List the UNGA-DM datasets}
\usage{
list_datasets(session)
}
\arguments{
\item{session}{An object of class \code{ungadm_session} created by
\code{authenticate()}.}
}
\value{
This function returns a string vector containing the names of the
  datasets in the UNGA-DM Database.
}
\description{
This function lists all of the datasets that are in the UNGA-DM
  Database. A number of functions in the \code{ungadm} package have a
  \code{dataset} argument. The values returned by \code{list_datasets()} are
  valid for this argument.
}
\examples{
\dontrun{
session <- authenticate(
  username = "USERNAME",
  password = "PASSWORD"
)

out <- list_datasets(
  session = session
)}

}
