% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/list_functions.R
\name{list_variables}
\alias{list_variables}
\title{List variables in a UNGA-DM dataset}
\usage{
list_variables(session, dataset)
}
\arguments{
\item{session}{An object of class \code{ungadm_session} created by
\code{authenticate()}.}

\item{dataset}{A string. The name of a dataset in the UNGA-DM Database Run
\code{list_datasets()} to get a list of valid values.}
}
\value{
This function returns a string vector containing the names of the
  variables in the specified dataset.
}
\description{
This function lists all of the variables in a dataset in the
  UNGA-DM Database. You have to specify a dataset.
}
\examples{
\dontrun{
session <- authenticate(
  username = "USERNAME",
  password = "PASSWORD"
)

out <- list_variables(
  session = session,
  dataset = "decisions"
)}

}
