% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print_citation.R
\name{print_citation}
\alias{print_citation}
\title{Print the citations for the UNGA-DM Database}
\usage{
print_citation(version)
}
\arguments{
\item{version}{A string. A version number in the format \code{#.#}, such as
\code{0.1}.}
}
\description{
This function generates the recommended citations for the UNGA-DM Database.
You have to provide a version number.
}
\examples{
\dontrun{
print_citation(version = "0.1")}

}
